bit of refactoring, adding status command
