+++ bin/losant-datatables.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 require('../lib/on-death');
 const unknownCommandWrap = require('../lib/unknown-command-listener');
-unknownCommandWrap(require('../commands/data-tables')).parse(process.argv);
+unknownCommandWrap(require('../commands/dataTables')).parse(process.argv);
