+++ commands/dataTables/status.js
@@ -1,0 +1,5 @@
+const { experience } = require('../../lib/constants');
+
+module.exports = (program) => {
+  return require('../utils/status')(program, 'dataTables', experience);
+};
