--- dataTables/Big Data.csv
@@ -1,2 +1,0 @@
-uh,yuh,id,updatedAt,createdAt
-hell,yeah,5ce826f2dfdcd00007f7d10e,2019-05-24T17:16:34.807Z,2019-05-24T17:16:34.807Z
