--- dataTables/Bigger Data.csv
@@ -1,2 +1,0 @@
-Big,Number,id,updatedAt,createdAt
-uh,80085,5ce8271fc9d9c8000aaa6ee3,2019-05-24T17:17:19.615Z,2019-05-24T17:17:19.615Z
